mods_trace: CPORTs are 16-bits not 8-bits
Switch to 16-bits for the CPORTs so we can see those in the trace -
specificaly the vendor control ports.

Reviewed-on: http://gerrit.mot.com/800247
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted

+++ mods_trace.h
@@ -26,7 +26,7 @@ struct gb_operation_msg_hdr;
 TRACE_EVENT(mods_switch,
 
 	TP_PROTO(struct gb_operation_msg_hdr *hdr, u8 from_intf,
-		u8 from_cport, u8 to_intf, u8 to_cport),
+		u16 from_cport, u8 to_intf, u16 to_cport),
 
 	TP_ARGS(hdr, from_intf, from_cport, to_intf, to_cport),
